refactor(data-loader): WIP extract attachments downloading procedure
